Exctract suggestion into new class to help with ruby critic.
